# openapi/evry-v1.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/auth/register": {
      "post": {
        "operationId": "webRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "headers": {
              "Set-Cookie": {
                "description": "Sets the HttpOnly evry_refresh browser session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "The email is already registered.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "webLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "headers": {
              "Set-Cookie": {
                "description": "Sets the HttpOnly evry_refresh browser session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [],
        "summary": "",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotates the refresh token stored in the HttpOnly evry_refresh cookie.",
        "operationId": "webRefresh",
        "parameters": [],
        "responses": {
          "200": {
            "headers": {
              "Set-Cookie": {
                "description": "Sets the HttpOnly evry_refresh browser session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or invalid browser session.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshCookie": []
          }
        ],
        "summary": "",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Revokes the cookie session when present and always clears the browser refresh cookie.",
        "operationId": "webLogout",
        "parameters": [],
        "responses": {
          "200": {
            "headers": {
              "Set-Cookie": {
                "description": "Clears the evry_refresh browser session cookie.",
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponse"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "refreshCookie": []
          },
          {}
        ],
        "summary": "",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/me": {
      "get": {
        "operationId": "authenticatedUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthUser"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "A valid access token is required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "currentUser",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The user profile no longer exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Users"
        ]
      },
      "patch": {
        "operationId": "updateCurrentUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdatedUser"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "The user profile no longer exists.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/adaptive/recommend/{exerciseId}": {
      "get": {
        "operationId": "AdaptiveController_recommend",
        "parameters": [
          {
            "name": "exerciseId",
            "required": true,
            "in": "path",
            "description": "Identificador del ejercicio para consultar el historial comparable.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Recomendación de carga y repeticiones basada en sesiones completadas y estado diario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdaptiveRecommendation"
                }
              }
            }
          },
          "400": {
            "description": "Solicitud inválida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Autenticación requerida.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Límite de solicitudes alcanzado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error interno normalizado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "tags": [
          "adaptive"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EVRY API",
    "description": "API privada de seguimiento de fuerza EVRY",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "evry_refresh"
      }
    },
    "schemas": {
      "RegisterInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "description": "Trimmed and lowercased before registration."
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "writeOnly": true
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 100,
            "description": "Leading and trailing whitespace is removed before validation."
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string",
            "nullable": true,
            "default": "PREFER_NOT_SAY",
            "description": "Omitted or null values use PREFER_NOT_SAY."
          },
          "trackCycle": {
            "type": "boolean",
            "nullable": true,
            "default": false,
            "description": "Omitted or null values disable cycle tracking."
          }
        },
        "required": [
          "email",
          "password",
          "name"
        ]
      },
      "AccessToken": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "LoginInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "description": "Trimmed and lowercased before authentication."
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LogoutResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "trackCycle": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "biologicalSex",
          "trackCycle"
        ]
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer"
          },
          "avgPeriodLen": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "biologicalSex",
          "birthDate",
          "goals",
          "trackCycle",
          "avgCycleLen",
          "avgPeriodLen",
          "createdAt"
        ]
      },
      "UserUpdateInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "description": "ISO 8601 date or date-time. Omitted or null values leave the current birth date unchanged."
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer",
            "minimum": 20,
            "maximum": 45
          },
          "avgPeriodLen": {
            "type": "integer",
            "minimum": 2,
            "maximum": 10
          }
        }
      },
      "UpdatedUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "name": {
            "type": "string"
          },
          "biologicalSex": {
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER",
              "PREFER_NOT_SAY"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          },
          "goals": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "STRENGTH",
                "HYPERTROPHY",
                "ENDURANCE",
                "FAT_LOSS",
                "GENERAL_FITNESS",
                "MOBILITY"
              ]
            }
          },
          "trackCycle": {
            "type": "boolean"
          },
          "avgCycleLen": {
            "type": "integer"
          },
          "avgPeriodLen": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "biologicalSex",
          "birthDate",
          "goals",
          "trackCycle",
          "avgCycleLen",
          "avgPeriodLen"
        ]
      },
...
      "AdaptiveRecommendation": {
        "type": "object",
        "properties": {
          "exerciseId": {
            "type": "string"
          },
          "targetWeightKg": {
            "type": "number",
            "nullable": true
          },
          "targetReps": {
            "type": "integer",
            "nullable": true
          },
          "rationale": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "action": {
            "enum": [
              "PROGRESS",
              "HOLD",
              "DELOAD"
            ],
            "type": "string",
            "description": "Acción emitida por el recomendador; HOLD también se usa cuando falta historial comparable."
          }
        },
        "required": [
          "exerciseId",
          "targetWeightKg",
          "targetReps",
          "rationale",
          "confidence",
          "action"
        ]
      },
...
      "ApiError": {
        "type": "object",
        "required": [
          "code",
          "message",
          "retryable",
          "requestId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "retryable": {
            "type": "boolean"
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}